feat(#9) :: room 도메인 rooms 이름변경 및 rooms 엔티티에 imageUrl 열 추가
