--- src/room/room.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { RoomService } from './room.service';
-import { RoomController } from './room.controller';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Room } from '@/src/room/entities/room.entity';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Room])],
-  controllers: [RoomController],
-  providers: [RoomService],
-})
-export class RoomModule {}
