--- src/room/room.service.ts
@@ -1,26 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { CreateRoomDto } from './dto/create-room.dto';
-import { UpdateRoomDto } from './dto/update-room.dto';
-
-@Injectable()
-export class RoomService {
-  create(createRoomDto: CreateRoomDto) {
-    return 'This action adds a new room';
-  }
-
-  findAll() {
-    return `This action returns all room`;
-  }
-
-  findOne(id: number) {
-    return `This action returns a #${id} room`;
-  }
-
-  update(id: number, updateRoomDto: UpdateRoomDto) {
-    return `This action updates a #${id} room`;
-  }
-
-  remove(id: number) {
-    return `This action removes a #${id} room`;
-  }
-}
