+++ src/rooms/dto/create-room.dto.ts
