+++ src/rooms/dto/update-room.dto.ts
