+++ src/rooms/rooms.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { RoomController } from './room.controller';
-import { RoomService } from './room.service';
+import { RoomController } from './rooms.controller';
+import { RoomsService } from './rooms.service';
 
 describe('RoomController', () => {
   let controller: RoomController;
@@ -8,7 +8,7 @@ describe('RoomController', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [RoomController],
-      providers: [RoomService],
+      providers: [RoomsService],
     }).compile();
 
     controller = module.get<RoomController>(RoomController);
