+++ src/rooms/rooms.controller.ts
@@ -1,0 +1,7 @@
+import { Controller } from '@nestjs/common';
+import { RoomsService } from './rooms.service';
+
+@Controller('room')
+export class RoomController {
+  constructor(private readonly roomService: RoomsService) {}
+}
